Add acceptance tests for mc NAT IP default rule
Since an MC CR needs to be created now (and it's expected to be a
single on on the acceptance cluster) the tests now run synchronously
and are in a single It. This is because there can be only one router
created as well. This can potentially be fixed with adding a
SynchronisedBeforeSuite which creates these resources and the MC, but
it would also make the tests ugly without also refactoring and adding
some kind of Fixture struct.