Add default security policy rules (#25)
* Extract common functions

* Add rule allowing MC NAT IPs to access WC api

* Implement NAT IP Resolver

* Add acceptance tests for mc NAT IP default rule

Since an MC CR needs to be created now (and it's expected to be a
single on on the acceptance cluster) the tests now run synchronously
and are in a single It. This is because there can be only one router
created as well. This can potentially be fixed with adding a
SynchronisedBeforeSuite which creates these resources and the MC, but
it would also make the tests ugly without also refactoring and adding
some kind of Fixture struct.

* Limit unit and acceptance tests to 8 nodes

Performance benefit from more nodes is negligable, however a lot more
resources need to be created (envtest and gcp resources) which might be
overwhelming. In ci it runs the tests with 35 nodes.

* Use IsNilOrEmpty consistently

* Extract security policy reconciler from controller

* Extract FirewallReconciler from controller

* Add default allow list for the Kube API

Currently by default this is the Giant Swarm VPN CIDrs

* Add default allow list for the Bastion hosts

Currently by default this is the Giant Swarm VPN CIDrs

* Address feedback

* log missing annotation
* add function for firewall tag

* Check if Router is set before reconciling

* Revert 'Check if Router is set before reconciling'

Reverts 74210b358e5fa8883a6e653d7ba8469d52973482